+++ CloudShell.ResourceModel/ResourceResolver.cs
@@ -1022,6 +1022,7 @@ public static class ResourceDefinitionDiagnosticCodes
     public const string ResourceChangeApplyProviderMissing = "resourceDefinition.changeApplyProviderMissing";
     public const string ResourceGraphVersionConflict = "resourceDefinition.graphVersionConflict";
     public const string ResourceGraphResourceMissing = "resourceDefinition.graphResourceMissing";
+    public const string ResourceGraphResourceAlreadyExists = "resourceDefinition.graphResourceAlreadyExists";
     public const string ResourceDefinitionTargetAmbiguous = "resourceDefinition.targetAmbiguous";
     public const string ResourceDefinitionIdentityChangeNotAllowed = "resourceDefinition.identityChangeNotAllowed";
     public const string ResourceDependencyCycle = "resourceDefinition.dependencyCycle";